+++ base/boot.jl
@@ -147,7 +147,7 @@ export
     # constants
     nothing, Main
 
-typealias AnyVector Array{Any,1}
+const AnyVector = Array{Any,1}
 
 abstract type Number end
 abstract type Real     <: Number end
@@ -175,9 +175,9 @@ primitive type Int128  <: Signed   128 end
 primitive type UInt128 <: Unsigned 128 end
 
 if Int === Int64
-    typealias UInt UInt64
+    const UInt = UInt64
 else
-    typealias UInt UInt32
+    const UInt = UInt32
 end
 
 function Typeof end
@@ -298,7 +298,7 @@ convert{T}(::Type{T}, x::T) = x
 cconvert{T}(::Type{T}, x) = convert(T, x)
 unsafe_convert{T}(::Type{T}, x::T) = x
 
-typealias NTuple{N,T} Tuple{Vararg{T,N}}
+NTuple{N,T} = Tuple{Vararg{T,N}}
 
 
 # primitive array constructors