+++ base/sparse/abstractsparse.jl
@@ -2,8 +2,8 @@
 
 abstract type AbstractSparseArray{Tv,Ti,N} <: AbstractArray{Tv,N} end
 
-typealias AbstractSparseVector{Tv,Ti} AbstractSparseArray{Tv,Ti,1}
-typealias AbstractSparseMatrix{Tv,Ti} AbstractSparseArray{Tv,Ti,2}
+AbstractSparseVector{Tv,Ti} = AbstractSparseArray{Tv,Ti,1}
+AbstractSparseMatrix{Tv,Ti} = AbstractSparseArray{Tv,Ti,2}
 
 """
     issparse(S)